+++ Source/Debug/Scratchpad.cpp
@@ -27,15 +27,15 @@ namespace Silent::Debug
     Vector2 SnapToGrid(const Vector2& pos, const Vector2& res, float virtualHeight)
     {
         float rawScale = res.y / virtualHeight;
-        float scale    = std::max(1.0f, std::floor(rawScale));
+        float scale    = std::max(1.0f, std::floorf(rawScale));
 
         //auto gridSize = Vector2(res.x / scale, virtualHeight); // Precise horizontal scaling.
         auto gridSize = Vector2(res.x / rawScale, virtualHeight);
 
         auto uv = pos / 100.0f;
 
-        auto snappedPixel = Vector2(std::floor(uv.x * gridSize.x), 
-                                    std::floor(uv.y * gridSize.y));
+        auto snappedPixel = Vector2(std::floorf(uv.x * gridSize.x), 
+                                    std::floorf(uv.y * gridSize.y));
 
         return Vector2((snappedPixel.x / gridSize.x) * 100.0f,
                        (snappedPixel.y / gridSize.y) * 100.0f);
